refactor : 베팅 금액 검증 책임 이동
